Move country fetch to provider for usable components, and fetch within widget
