+++ app/javascript/pages/country/root/root-reducers.js
@@ -1,60 +1,9 @@
 export const initialState = {
-  isCountriesLoading: false,
-  isRegionsLoading: false,
-  isSubRegionsLoading: false,
-  isGeostoreLoading: false,
   gfwHeaderHeight: 59,
   showMapMobile: false,
   isMapFixed: true,
-  mapTop: 0,
-  countries: [],
-  regions: [],
-  subRegions: [],
-  geostore: {
-    areaHa: 0,
-    bounds: []
-  }
+  mapTop: 0
 };
-
-const setCountriesLoading = (state, { payload }) => ({
-  ...state,
-  isCountriesLoading: payload
-});
-
-const setRegionsLoading = (state, { payload }) => ({
-  ...state,
-  isRegionsLoading: payload
-});
-
-const setSubRegionsLoading = (state, { payload }) => ({
-  ...state,
-  isSubRegionsLoading: payload
-});
-
-const setGeostoreLoading = (state, { payload }) => ({
-  ...state,
-  isGeostoreLoading: payload
-});
-
-const setCountries = (state, { payload }) => ({
-  ...state,
-  countries: payload.map(d => ({ label: d.name, value: d.iso }))
-});
-
-const setRegions = (state, { payload }) => ({
-  ...state,
-  regions: payload.map(d => ({ label: d.name, value: d.id }))
-});
-
-const setSubRegions = (state, { payload }) => ({
-  ...state,
-  subRegions: payload.map(d => ({ label: d.name, value: d.id }))
-});
-
-const setGeostore = (state, { payload }) => ({
-  ...state,
-  geostore: payload
-});
 
 const setFixedMapStatus = (state, { payload }) => ({
   ...state,
@@ -72,14 +21,6 @@ const setShowMapMobile = (state, { payload }) => ({
 });
 
 export default {
-  setCountriesLoading,
-  setRegionsLoading,
-  setSubRegionsLoading,
-  setGeostoreLoading,
-  setCountries,
-  setRegions,
-  setSubRegions,
-  setGeostore,
   setFixedMapStatus,
   setMapTop,
   setShowMapMobile
